+++ create_graph.h
@@ -11,6 +11,6 @@ struct Graph* constructGraphFromFile(const string& strFileName);
 
 struct Graph* constructGraphFromUserInput();
 
-vector<tuple<int, int, float, int>> constructAddressFromFile(const string& strFileName);
+vector<tuple<int, int, double, int>> constructAddressFromFile(const string& strFileName);
 
-vector<tuple<int, int, float, int>> constructAddressFromUserInput();
+vector<tuple<int, int, double, int>> constructAddressFromUserInput();
